Doing some reorganizing before implementing random search.
