--- bayesian_optimizers/bayesian_optimizer.py
@@ -1,8 +1,0 @@
-from abc import ABC, abstractmethod
-
-from util.defs import *
-
-class BayesianOptimizer(ABC):
-    @abstractmethod
-    def optimize(self, iterations, objective) -> Dict:
-        pass
